+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Name changed from 'stac-api' to 'stac-server' to avoid confusion with [stac-api-spec](https://github.com/radiantearth/stac-api-spec)
 - Elasticsearch client library updated to @elastic/elasticsearch
 
+### Removed
+- Collection properties no longer merged into Item on ingest (commons extension removed from STAC)
+
 
 ## [0.1.0] - 2020-03-20